modify code according feedback

+++ include/rapidjson/internal/strtod.h
@@ -135,7 +135,7 @@ inline bool StrtodDiyFp(const char* decimals, int dLen, int dExp, double* result
         if (significand  >  RAPIDJSON_UINT64_C2(0x19999999, 0x99999999) ||
             (significand == RAPIDJSON_UINT64_C2(0x19999999, 0x99999999) && decimals[i] > '5'))
             break;
-        significand = significand * 10u + decimals[i] - '0';
+        significand = significand * 10u + static_cast<unsigned>(decimals[i]) - '0';
     }
     
     if (i < dLen && decimals[i] >= '5') // Rounding
